+++ src/main/java/groupware/PlaygroundApplication.java
@@ -1,4 +1,4 @@
-package playground;
+package groupware;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
